Add files via upload
This is the a customer care chat-bot specifically for banking systems. This bot is actually working on OPENAI's GPT-3 model and given a saved file which contains various queries regarding banks. The bot contain two of the models to perform specific task to give the answer to the input query.